audio thread done kinda right and audio sync
audio thread is unlocked at the end of the frame and execute the remaining cycles after the lazy updates that I have yet to implement, audio is now synced based on frametime and the fps limiter is also better. Also changed some types

+++ src/core/bus.h
@@ -47,5 +47,5 @@ private:
 	bool m_vramBusBlocked;
 	uint16 m_dmaTransferCurrentAddress;
 	bool m_dmaTransferInProcess;
-	int m_dmaTransferEnableDelay;
+	uint8 m_dmaTransferEnableDelay;
 };
